Move assert_err into test_util
