+++ src/http/content_type.rs
@@ -172,7 +172,7 @@ impl ContentType {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{test_util::_Factori_Builder_Response, util::assert_err};
+    use crate::test_util::*;
     use factori::create;
     use reqwest::header::{
         HeaderMap, HeaderValue, InvalidHeaderValue, CONTENT_TYPE,
