+++ src/test_util.rs
@@ -186,3 +186,21 @@ pub fn header_map<'a>(
         })
         .collect()
 }
+
+/// Assert a result is the `Err` variant, and the stringified error contains
+/// the given message
+macro_rules! assert_err {
+    ($e:expr, $msg:expr) => {{
+        use itertools::Itertools as _;
+
+        let msg = $msg;
+        // Include all source errors so wrappers don't hide the important stuff
+        let error: anyhow::Error = $e.unwrap_err().into();
+        let actual = error.chain().map(ToString::to_string).join(": ");
+        assert!(
+            actual.contains(msg),
+            "Expected error message to contain {msg:?}, but was: {actual:?}"
+        )
+    }};
+}
+pub(crate) use assert_err;
