+++ src/util.rs
@@ -303,24 +303,6 @@ where
 }
 
 #[cfg(test)]
-macro_rules! assert_err {
-    ($e:expr, $msg:expr) => {{
-        use itertools::Itertools as _;
-
-        let msg = $msg;
-        // Include all source errors so wrappers don't hide the important stuff
-        let error: anyhow::Error = $e.unwrap_err().into();
-        let actual = error.chain().map(ToString::to_string).join(": ");
-        assert!(
-            actual.contains(msg),
-            "Expected error message to contain {msg:?}, but was: {actual:?}"
-        )
-    }};
-}
-#[cfg(test)]
-pub(crate) use assert_err;
-
-#[cfg(test)]
 mod tests {
     use super::*;
     use strum::EnumIter;
